backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
ruff
